requirements.txt
tqdm~=4.64.0
azure-identity
azure-search-documents
joblib~=1.4.2
streamlink
aiohttp

requests~=2.31.0